+ do nothing when no main.js

+++ generator/helpers.js
@@ -4,7 +4,6 @@ module.exports = function (api) {
   return {
     updateBabelConfig (callback) {
       let config, configPath
-
       const rcPath = api.resolve('./babel.config.js')
       const pkgPath = api.resolve('./package.json')
       if (fs.existsSync(rcPath)) {
@@ -13,17 +12,14 @@ module.exports = function (api) {
       } else if (fs.existsSync(pkgPath)) {
         configPath = pkgPath
         config = JSON.parse(fs.readFileSync(pkgPath, { encoding: 'utf8' }))
-
         if (config.babel) {
           config.babel = callback(config.babel)
         } else {
           // TODO: error handling here?
         }
       }
-
       if (configPath) {
         const moduleExports = configPath !== pkgPath ? 'module.exports = ' : ''
-
         fs.writeFileSync(
           configPath,
           `${moduleExports}${JSON.stringify(config, null, 2)}`,
@@ -37,14 +33,11 @@ module.exports = function (api) {
     updateMain (callback) {
       const tsPath = api.resolve('./src/main.ts')
       const jsPath = api.resolve('./src/main.js')
-
       const mainPath = fs.existsSync(tsPath) ? tsPath : jsPath
+      if (!fs.existsSync(mainPath)) return
       let content = fs.readFileSync(mainPath, { encoding: 'utf8' })
-
       let lines = content.split(/\r?\n/g)
-
       lines = callback(lines)
-
       content = lines.join('\n')
       fs.writeFileSync(mainPath, content, { encoding: 'utf8' })
     }
